requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
arrow==0.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
asgiref==3.3.1; python_version >= '3.5'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
automat==20.2.0
beautifulsoup4==4.9.3
blessed==1.17.12
boto3==1.17.12
botocore==1.20.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
configargparse==1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
defusedxml==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-rest-auth==2.1.3
django-admin-object-actions==0.1.11
django-admin-rangefilter==0.6.4
django-admin-relation-links==0.2.5
django-allauth==0.44.0
django-anymail[amazon_ses]==8.2
django-ckeditor==6.0.0
django-clearcache==1.1.1
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-crum==0.7.9
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.4.0
django-js-asset==1.2.2
django-json-widget==1.1.1
django-modern-rpc==0.12.1
django-nested-admin==3.3.3
django-nested-inline==0.4.2
django-parler==2.2
django-picklefield==3.0.1; python_version >= '3'
django-q==1.3.4
django-redis==4.12.1
django-rest-passwordreset==1.1.0
django-rq==2.4.0
django-storages==1.11.1
django-timezone-field==4.1.1
django-uuslug==1.2.0
django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
flask-basicauth==0.2.0
flask==1.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gevent==21.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
geventhttpclient==1.4.4
greenlet==1.0.0; platform_python_implementation == 'CPython'
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
incremental==17.5.0
inflection==0.5.1; python_version >= '3.5'
itemadapter==0.2.0; python_version >= '3.6'
itemloaders==1.0.4; python_version >= '3.6'
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itypes==1.2.0
jinja2==2.11.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
locust==1.4.3
lxml==4.6.2; platform_python_implementation == 'CPython'
markdown==3.3.3
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
msgpack==1.0.2
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parsel==1.6.0
protego==0.1.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
psutil==5.8.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
psycopg2-binary==2.8.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydispatcher==2.0.5
pyhamcrest==2.0.2; python_version >= '3.5'
pyjwt[crypto]==2.0.1; python_version >= '3.6'
pyopenssl==20.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pypinyin==0.40.0
pysocks==1.7.1
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-monkey-business==1.0.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pyzmq==22.0.3; python_version >= '3.6'
queuelib==1.5.0
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
requests-oauthlib==1.3.0
requests[socks]==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rq==1.7.0; python_version >= '3.5'
ruamel.yaml.clib==0.2.2; python_version < '3.9' and platform_python_implementation == 'CPython'
ruamel.yaml==0.16.12
s3transfer==0.3.4
scrapy==2.4.1
service-identity==18.1.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2; python_version >= '3'
sqlparse==0.4.1; python_version >= '3.5'
text-unidecode==1.3
tweepy==3.10.0
twisted==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
ujson==4.0.2
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
w3lib==1.22.0
wcwidth==0.2.5
werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
